Refactor: Refactor connection functions
